Fix CI [WinError 5] Access is denied
Windows doesn't allow changing the running executable. https://github.com/pypa/pip/issues/ 1299#issuecomment-27809600
It worked for so long because earlier pip versions provided by azure
were already the latest ones and `--upgrade` did not modify the running
pip. Now, suddenly a new pip was released which was not present in Azure
windows and the `--upgrade pip` was triggered modifying the very
executable it is running.
Why does it say "Access is denied" (a permission error) is a mystery yet
to be unrevealed by Microsoft folks.

That said, untouched, this bug would have disappeared (when Azure
updates the pip version) and reappeared again and again haunting us for
long. No more.

Signed-off-by: Hritik Vijay <hritikxx8@gmail.com>